[NFC] Move areCompatibleSveTypes etc. from ASTContext to SemaARM.
In preparation for making these functions interact with the current
context; see #144611.
